backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
